+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 > Released: 2026-04-xx
 
 - [Added] Timings to Functions
-- [Added] New-PartialVVFBringupValidation and New-PartialVVFBringup functions
+- [Added] New-VVFBasedPartialBringupValidation and New-VVFBasedPartialBringup functions
 - [Changed] Handle multiple ip address retrieval during New-ExtractDataFromSDDCBackup
 
 ## v9.0.1